Put favorites list for dashboard into its own folder and renamed ChartFavorites to DashboardFavorites
