+++ src/components/dashboard/dbFavorites/FavoriteSlots.vue
